Line 103: setItems(result.items);
I copied and pasted the React docs solution for fetching an API with React Hooks, and everything went well except for line 103, but the error returned happened at line 133.

If 103:
```
setItems(result);
```
Then the error is 'items.map()' is not a function. 

When I changed the line 103 to setItems(result.items) everything works fine.